requirements.txt
aws-cdk-lib==2.164.1
constructs>=10.0.0,<11.0.0
boto3>=1.26.0
requests>=2.31.0
